Rename training pipeline directory
