+++ src/genai_detection/pipelines/training/pipeline.py
@@ -1,5 +1,5 @@
 from kedro.pipeline import Pipeline, pipeline
-from genai_detection.pipelines.data_processing.nodes import (
+from genai_detection.pipelines.training.nodes import (
     preprocess_raw_image_data_node,
     model_initialization_node,
     train_model_node,
